[REFACTOR] Arguments namespace does not makes sense.
+ Renames argument namespace into options.

+ Solve ARM already identified (occurs when a switch expression have duplicate conditions).

(null, null) is the same as (var x, null) but not the same as (string x, null).

var on type patterns means any type that is fetched from a given variable and then it pass it's value by reference for refs and value for valuetype to a given variable.

Thus you can't do var x  = null; impossible to infer the type at compile time. you can make
var x = 9; x is var y and then y is 9 but also int? number = null!; number is var n.  And then n gets nulled. Cool <3 type patterns, that's the expected behavior due to var meaning any type and null is only a value that means absense of reference to an object on heap.